requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
redis==5.0.4
boto3==1.34.100
razorpay==1.4.1
httpx==0.27.0
anthropic==0.28.0
openpyxl==3.1.2
gspread==6.1.2
google-auth==2.29.0
